use size_t as GetStateSize return optimize performance in agg getoutput and fix sum decimal agg clean code & fix review & fix bug : no need to care overflow when sum flat im

+++ core/src/operator/aggregation/aggregator/aggregator.h
@@ -239,7 +239,7 @@ public:
         aggStateOffset = offset;
     }
 
-    virtual int32_t GetStateSize() = 0;
+    virtual size_t GetStateSize() = 0;
 
     virtual void InitStates(std::vector<AggregateState *> &groupStates)
     {
