+++ core/src/operator/aggregation/aggregator/average_spark_decimal_aggregator.h
@@ -75,7 +75,7 @@ public:
         }
     }
 
-    int32_t GetStateSize() override
+    size_t GetStateSize() override
     {
         return sizeof(AvgSparkDecimalState);
     }
