+++ core/src/operator/aggregation/aggregator/first_aggregator.h
@@ -414,7 +414,7 @@ public:
         }
     }
 
-    int32_t GetStateSize() override
+    size_t GetStateSize() override
     {
         return sizeof(FirstState);
     }
