+++ core/src/operator/aggregation/aggregator/min_varchar_aggregator.h
@@ -311,7 +311,7 @@ public:
         return spillTypes;
     }
 
-    int32_t GetStateSize() override
+    size_t GetStateSize() override
     {
         return sizeof(VarcharState);
     }
