+++ core/src/operator/aggregation/aggregator/typed_mask_column_assistant_aggregator.h
@@ -213,7 +213,7 @@ public:
         realAggregator->SetStateOffset(offset);
     }
 
-    int32_t GetStateSize() override
+    size_t GetStateSize() override
     {
         return realAggregator->GetStateSize();
     }
